feat: Android v1.2.0 Release 问题清单

+++ AlcedoAndroid/app/src/main/java/com/alcedo/studio/domain/service/MaskRenderService.kt
@@ -82,7 +82,7 @@ class MaskRenderService(
     suspend fun renderMaskOverlay(
         containers: List<MaskContainer>,
         bitmap: Bitmap,
-        argbColor: Int = 0xCCFF3030
+        argbColor: Int = 0xCCFF3030.toInt()
     ): Bitmap = withContext(Dispatchers.Default) {
         val w = bitmap.width
         val h = bitmap.height
